refactoring jwt utils

+++ internal/models/Course.go
@@ -2,7 +2,7 @@ package models
 
 type Course struct {
 	Model
-	UserId      uint64 `json:"user_id"`
+	UserId      int    `json:"user_id"`
 	Title       string `json:"title" validate:"required"`
 	Description string `json:"description" validate:"required"`
 }
