+++ internal/models/Model.go
@@ -9,7 +9,7 @@ import (
 )
 
 type Model struct {
-	ID        uint64         `json:"id" gorm:"primarykey"`
+	ID        int            `json:"id" gorm:"primarykey"`
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `json:"deleted_at" gorm:"index"`
